# server: Implement non-blocking IO for VR bridges