--- a/server/src/main/java/dev/slimevr/VRServer.java
+++ b/server/src/main/java/dev/slimevr/VRServer.java
@@ -295,7 +295,9 @@ public class VRServer extends Thread {
 			// final long time = System.currentTimeMillis() - start;
 			try {
 				Thread.sleep(1); // 1000Hz
-			} catch (InterruptedException ignored) {}
+			} catch (InterruptedException ignored) {
+				// waked from WindowsNamedPipeBridge, UnixSocketBridge
+			}
 		}
 	}