--- a/server/src/main/java/dev/slimevr/VRServer.java
+++ b/server/src/main/java/dev/slimevr/VRServer.java
@@ -66,6 +66,7 @@ public class VRServer extends Thread {
 	private final NanoTimer fpsTimer = new NanoTimer();
 	private final ProvisioningHandler provisioningHandler;
 	private final ResetHandler resetHandler;
+	private final Object wakeupLock = new Object();
 
 	/**
 	 * This function is used by VRWorkout, do not remove!
@@ -297,11 +298,18 @@ public class VRServer extends Thread {
 			vrcOSCHandler.update();
 			vmcHandler.update();
 			// final long time = System.currentTimeMillis() - start;
-			try {
-				Thread.sleep(1); // 1000Hz
-			} catch (InterruptedException ignored) {
-				// waked from WindowsNamedPipeBridge, UnixSocketBridge
+			synchronized (wakeupLock) {
+				try {
+					wakeupLock.wait(1); // 1000Hz
+				} catch (InterruptedException ignored) {}
 			}
+		}
+	}
+
+	@ThreadSafe
+	public void wakeup() {
+		synchronized (wakeupLock) {
+			wakeupLock.notify();
 		}
 	}