--- a/server/src/main/java/dev/slimevr/bridge/ProtobufBridge.java
+++ b/server/src/main/java/dev/slimevr/bridge/ProtobufBridge.java
@@ -56,11 +56,7 @@ public abstract class ProtobufBridge<T extends VRTracker> implements Bridge {
 	 */
 	@ThreadSafe
 	protected void signalRecv() {
-		try {
-			Main.getVrServer().interrupt();
-		} catch (SecurityException ex) {
-			// do nothing
-		}
+		Main.getVrServer().wakeup();
 	}
 
 	@BridgeThread