--- a/server/src/main/java/dev/slimevr/platform/windows/WindowsNamedPipeBridge.java
+++ b/server/src/main/java/dev/slimevr/platform/windows/WindowsNamedPipeBridge.java
@@ -1,11 +1,12 @@
 package dev.slimevr.platform.windows;
 
 import com.google.protobuf.CodedOutputStream;
+import com.google.protobuf.InvalidProtocolBufferException;
+import com.sun.jna.Native;
+import com.sun.jna.platform.win32.*;
 import com.sun.jna.platform.win32.Kernel32;
-import com.sun.jna.platform.win32.WinBase;
-import com.sun.jna.platform.win32.WinError;
-import com.sun.jna.platform.win32.WinNT;
 import com.sun.jna.ptr.IntByReference;
+import com.sun.jna.win32.W32APIOptions;
 import dev.slimevr.Main;
 import dev.slimevr.VRServer;
 import dev.slimevr.bridge.BridgeThread;
@@ -20,15 +21,37 @@ import java.io.IOException;
 import java.util.List;
 
 
+interface Kernel32IO extends Kernel32 {
+	Kernel32IO INSTANCE = Native.load("kernel32", Kernel32IO.class, W32APIOptions.DEFAULT_OPTIONS);
+
+	boolean GetOverlappedResult(
+		/* [in] */ HANDLE hFile,
+		/* [in] */ WinBase.OVERLAPPED lpOverlapped,
+		/* [out] */ IntByReference lpNumberOfBytesTransferred,
+		/* [in] */ boolean bWait
+	);
+}
+
+
 public class WindowsNamedPipeBridge extends SteamVRBridge {
+	private static final Kernel32 k32 = Kernel32.INSTANCE;
+	private static final Kernel32IO k32io = Kernel32IO.INSTANCE;
 
 	protected final String pipeName;
 	private final byte[] buffArray = new byte[2048];
 	protected WindowsPipe pipe;
-	protected WinNT.HANDLE rxEvent = Kernel32.INSTANCE.CreateEvent(null, false, false, null);
-	protected WinNT.HANDLE txEvent = Kernel32.INSTANCE.CreateEvent(null, false, false, null);
+	protected WinNT.HANDLE readEvent = k32.CreateEvent(null, false, false, null);
+	protected WinNT.HANDLE writeEvent = k32.CreateEvent(null, false, false, null);
+	protected WinNT.HANDLE rxEvent = k32.CreateEvent(null, false, false, null);
+	protected WinNT.HANDLE txEvent = k32.CreateEvent(null, false, false, null);
 	protected WinNT.HANDLE[] events = new WinNT.HANDLE[] { rxEvent, txEvent };
-	protected WinBase.OVERLAPPED overlapped = new WinBase.OVERLAPPED();
+	WinBase.OVERLAPPED overlappedWrite = new WinBase.OVERLAPPED();
+	WinBase.OVERLAPPED overlappedRead = new WinBase.OVERLAPPED();
+	WinBase.OVERLAPPED overlappedWait = new WinBase.OVERLAPPED();
+	IntByReference bytesWritten = new IntByReference(0);
+	IntByReference bytesAvailable = new IntByReference(0);
+	IntByReference bytesRead = new IntByReference(0);
+	IntByReference bytesWait = new IntByReference(0);
 
 	public WindowsNamedPipeBridge(
 		VRServer server,
@@ -40,7 +63,6 @@ public class WindowsNamedPipeBridge extends SteamVRBridge {
 	) {
 		super(server, hmd, "Named pipe thread", bridgeName, bridgeSettingsKey, shareableTrackers);
 		this.pipeName = pipeName;
-		overlapped.hEvent = rxEvent;
 	}
 
 	@Override
@@ -81,107 +103,140 @@ public class WindowsNamedPipeBridge extends SteamVRBridge {
 	@Override
 	@ThreadSafe
 	protected void signalSend() {
-		Kernel32.INSTANCE.SetEvent(txEvent);
+		k32.SetEvent(txEvent);
 	}
 
 	@BridgeThread
 	private void waitForData(int timeoutMs) {
 		if (pipe.state != PipeState.OPEN)
 			return;
-		// async request for data, overlapped contains the rxEvent handle
-		Kernel32.INSTANCE.ReadFile(pipe.pipeHandle, null, 0, null, overlapped);
-		Kernel32.INSTANCE.WaitForMultipleObjects(events.length, events, false, timeoutMs);
+		if (k32io.GetOverlappedResult(pipe.pipeHandle, overlappedWait, bytesWait, false)) {
+			overlappedWait.clear();
+			overlappedWait.hEvent = rxEvent;
+		}
+		k32.ReadFile(pipe.pipeHandle, null, 0, null, overlappedWait);
+		k32.WaitForMultipleObjects(events.length, events, false, timeoutMs);
 	}
 
 	@Override
 	@BridgeThread
 	protected boolean sendMessageReal(ProtobufMessage message) {
-		if (pipe.state == PipeState.OPEN) {
-			try {
-				int size = message.getSerializedSize();
-				CodedOutputStream os = CodedOutputStream.newInstance(buffArray, 4, size);
-				message.writeTo(os);
-				size += 4;
-				buffArray[0] = (byte) (size & 0xFF);
-				buffArray[1] = (byte) ((size >> 8) & 0xFF);
-				buffArray[2] = (byte) ((size >> 16) & 0xFF);
-				buffArray[3] = (byte) ((size >> 24) & 0xFF);
-				if (Kernel32.INSTANCE.WriteFile(pipe.pipeHandle, buffArray, size, null, null)) {
-					return true;
-				}
-				pipe.state = PipeState.ERROR;
-				LogManager
-					.severe("[" + bridgeName + "] Pipe error: " + Kernel32.INSTANCE.GetLastError());
-			} catch (IOException e) {
-				e.printStackTrace();
+		if (pipe.state != PipeState.OPEN) {
+			return false;
+		}
+		try {
+			int size = message.getSerializedSize();
+			CodedOutputStream os = CodedOutputStream.newInstance(buffArray, 4, size);
+			message.writeTo(os);
+			size += 4;
+			buffArray[0] = (byte) (size & 0xFF);
+			buffArray[1] = (byte) ((size >> 8) & 0xFF);
+			buffArray[2] = (byte) ((size >> 16) & 0xFF);
+			buffArray[3] = (byte) ((size >> 24) & 0xFF);
+
+			overlappedWrite.clear();
+			overlappedWrite.hEvent = writeEvent;
+			boolean immediate = k32
+				.WriteFile(pipe.pipeHandle, buffArray, size, null, overlappedWrite);
+			int err = k32.GetLastError();
+			if (!immediate && err != WinError.ERROR_IO_PENDING) {
+				setPipeError("WriteFile failed: " + err);
+				return false;
 			}
+
+			bytesWritten.setValue(0);
+			if (!k32io.GetOverlappedResult(pipe.pipeHandle, overlappedWrite, bytesWritten, true)) {
+				setPipeError(
+					"sendMessageReal/GetOverlappedResult failed: " + k32.GetLastError()
+				);
+				return false;
+			}
+
+			if (bytesWritten.getValue() != size) {
+				setPipeError("Bytes written " + bytesWritten.getValue() + ", expected " + size);
+				return false;
+			}
+
+			return true;
+		} catch (IOException e) {
+			e.printStackTrace();
 		}
 		return false;
 	}
 
 	private boolean updatePipe() throws IOException {
-		if (pipe.state == PipeState.OPEN) {
-			boolean readAnything = false;
-			IntByReference bytesAvailable = new IntByReference(0);
-			while (
-				Kernel32.INSTANCE
-					.PeekNamedPipe(pipe.pipeHandle, buffArray, 4, null, bytesAvailable, null)
-			) {
-				if (bytesAvailable.getValue() >= 4) { // Got size
-					int messageLength = (buffArray[3] << 24)
-						| (buffArray[2] << 16)
-						| (buffArray[1] << 8)
-						| buffArray[0];
-					if (messageLength > 1024) { // Overflow
-						LogManager
-							.severe(
-								"["
-									+ bridgeName
-									+ "] Pipe overflow. Message length: "
-									+ messageLength
-							);
-						pipe.state = PipeState.ERROR;
-						return readAnything;
-					}
-					if (bytesAvailable.getValue() >= messageLength) {
-						if (
-							Kernel32.INSTANCE
-								.ReadFile(
-									pipe.pipeHandle,
-									buffArray,
-									messageLength,
-									bytesAvailable,
-									null
-								)
-						) {
-							ProtobufMessage message = ProtobufMessage
-								.parser()
-								.parseFrom(buffArray, 4, messageLength - 4);
-							messageReceived(message);
-							readAnything = true;
-						} else {
-							pipe.state = PipeState.ERROR;
-							LogManager
-								.severe(
-									"["
-										+ bridgeName
-										+ "] Pipe error: "
-										+ Kernel32.INSTANCE.GetLastError()
-								);
-							return readAnything;
-						}
-					} else {
-						return readAnything; // Wait for more data
-					}
-				} else {
-					return readAnything; // Wait for more data
-				}
-			}
-			pipe.state = PipeState.ERROR;
-			LogManager
-				.severe("[" + bridgeName + "] Pipe error: " + Kernel32.INSTANCE.GetLastError());
+		if (pipe.state != PipeState.OPEN) {
+			return false;
 		}
-		return false;
+		bytesAvailable.setValue(0);
+		boolean readAnything = false;
+		while (k32.PeekNamedPipe(pipe.pipeHandle, buffArray, 4, null, bytesAvailable, null)) {
+			if (bytesAvailable.getValue() < 4) {
+				return readAnything; // Wait for more data
+			}
+			int messageLength = (buffArray[3] << 24)
+				| (buffArray[2] << 16)
+				| (buffArray[1] << 8)
+				| buffArray[0];
+			if (messageLength > 1024) { // Overflow
+				setPipeError("Pipe overflow. Message length: " + messageLength);
+				return readAnything;
+			}
+			if (bytesAvailable.getValue() < messageLength) {
+				return readAnything; // Wait for more data
+			}
+
+			overlappedRead.clear();
+			overlappedRead.hEvent = readEvent;
+
+			boolean immediate = k32
+				.ReadFile(pipe.pipeHandle, buffArray, messageLength, null, overlappedRead);
+			int err = k32.GetLastError();
+			if (!immediate && err != WinError.ERROR_IO_PENDING) {
+				setPipeError("ReadFile failed: " + err);
+				return readAnything;
+			}
+
+			bytesRead.setValue(0);
+			if (!k32io.GetOverlappedResult(pipe.pipeHandle, overlappedRead, bytesRead, true)) {
+				setPipeError(
+					"updatePipe/GetOverlappedResult failed: " + k32.GetLastError()
+				);
+				return readAnything;
+			}
+
+			if (bytesRead.getValue() != messageLength) {
+				setPipeError(
+					"Bytes read " + bytesRead.getValue() + ", expected " + messageLength
+				);
+				return readAnything;
+			}
+
+			try {
+				ProtobufMessage message = ProtobufMessage
+					.parser()
+					.parseFrom(buffArray, 4, messageLength - 4);
+				messageReceived(message);
+				readAnything = true;
+			} catch (InvalidProtocolBufferException parseEx) {
+				parseEx.printStackTrace();
+				setPipeError("Failed to parse message: " + parseEx.getMessage());
+				return readAnything;
+			}
+		}
+
+		int err = k32.GetLastError();
+		if (err == WinError.ERROR_BROKEN_PIPE) {
+			setPipeError("Pipe closed");
+		} else {
+			setPipeError("Pipe error: " + err);
+		}
+		return readAnything;
+	}
+
+	private void setPipeError(String message) {
+		pipe.state = PipeState.ERROR;
+		LogManager.severe("[" + bridgeName + "] " + message);
 	}
 
 	private void resetPipe() {
@@ -193,7 +248,7 @@ public class WindowsNamedPipeBridge extends SteamVRBridge {
 	private void createPipe() throws IOException {
 		try {
 			pipe = new WindowsPipe(
-				Kernel32.INSTANCE
+				k32
 					.CreateNamedPipe(
 						pipeName,
 						WinBase.PIPE_ACCESS_DUPLEX | WinNT.FILE_FLAG_OVERLAPPED, // dwOpenMode
@@ -202,15 +257,14 @@ public class WindowsNamedPipeBridge extends SteamVRBridge {
 						1024 * 16, // nOutBufferSize,
 						1024 * 16, // nInBufferSize,
 						0, // nDefaultTimeOut,
-						null
+						null // lpSecurityAttributes
 					),
 				pipeName
-			); // lpSecurityAttributes
+			);
 			LogManager.info("[" + bridgeName + "] Pipe " + pipe.name + " created");
-			if (WinBase.INVALID_HANDLE_VALUE.equals(pipe.pipeHandle))
-				throw new IOException(
-					"Can't open " + pipeName + " pipe: " + Kernel32.INSTANCE.GetLastError()
-				);
+			if (WinBase.INVALID_HANDLE_VALUE.equals(pipe.pipeHandle)) {
+				throw new IOException("Can't open " + pipeName + " pipe: " + k32.GetLastError());
+			}
 			LogManager.info("[" + bridgeName + "] Pipes are created");
 		} catch (IOException e) {
 			WindowsPipe.safeDisconnect(pipe);
@@ -220,8 +274,8 @@ public class WindowsNamedPipeBridge extends SteamVRBridge {
 
 	private boolean tryOpeningPipe(WindowsPipe pipe) {
 		if (
-			Kernel32.INSTANCE.ConnectNamedPipe(pipe.pipeHandle, null)
-				|| Kernel32.INSTANCE.GetLastError() == WinError.ERROR_PIPE_CONNECTED
+			k32.ConnectNamedPipe(pipe.pipeHandle, null)
+				|| k32.GetLastError() == WinError.ERROR_PIPE_CONNECTED
 		) {
 			pipe.state = PipeState.OPEN;
 			LogManager.info("[" + bridgeName + "] Pipe " + pipe.name + " is open");
@@ -235,7 +289,7 @@ public class WindowsNamedPipeBridge extends SteamVRBridge {
 					+ "] Error connecting to pipe "
 					+ pipe.name
 					+ ": "
-					+ Kernel32.INSTANCE.GetLastError()
+					+ k32.GetLastError()
 			);
 		return false;
 	}