I forgot `FILE_FLAG_OVERLAPPED` and it was actually blocking at `ReadFile` on Windows.

Now it implements non-blocking IO for both receiving and sending, for the `UnixSocketBridge` too. It was tested on Linux, but only on synthetic data as I don't own a Linux-compatible VR headset. On Windows it felt slightly snappier and reduced hiccups, compared to the #607.

Latency became more consistent on Linux across different update rates. On Windows it's clear that it was receiving data with a delay of 1 frame.

Time delta measurements between sending headset position and receiving a new tracker position:

| Platform / Update rate / Num frames | Old (ms) | New (ms) |
| -------- | ---- | --- |
|                        | `   p1   avg    p99`  | `   p1   avg   p99` |
| Windows / 90hz / 5000  | `0.089 0.134  0.203`  | `0.538 1.746 4.103` |
| Windows / 100hz / 5000 | `0.086 0.130  0.213`  | `0.538 1.721 4.094` |
| Windows / 120hz / 5000 | `0.076 0.192  0.253`  | `0.550 1.844 4.123` |
| Linux / 90hz / 5000    | `0.117 1.685 11.799`  | `0.172 1.305 2.196` |
| Linux / 100hz / 5000   | `0.102 0.764 10.710`  | `0.157 1.086 1.815` |
| Linux / 120hz / 5000   | `0.087 0.622  8.701`  | `0.161 1.288 1.835` |

![image](https://user-images.githubusercontent.com/114709761/224910810-5f108ba0-e7a0-42b7-a639-a101f0d4c2c3.png)

Please review and test.