@0forks 2023-03-14
> what are you using to collect those statistics

100 lines of JS and plotly, it was faster to write a crossplatform tester in JS. Though it's always async and buffered (and slower than native), so doesn't represent the driver latency exactly. But should be close enough.

> the difference between windows old and windows

In the old version it's a delayed tracker frame, which comes instantly when the thread unblocks after receiving a headset frame. The new latency is what it should have been instead.

> I suspect this might still be handling some details w/ overlapped io incorrectly (i.e. we should be inserting explicit waits/getoverlappedresult anywhere we want a blocking-style call)

I'm calling `ReadFile` with 0 bytes to read, and `GetOverlappedResult` is only used to get amount of bytes read. `ReadFile` should always instantly return `false` with a `GetLastError` either 0 or IO pending. We're doing a select with `WaitForMultipleObjects` on auto-resetting events, so technically we should be fine here.

Indeed, there could be something I missed, so I want more people to review this to make sure all of it is implemented correctly, as I haven't touched Java and WinAPI in years.