@ButterscotchV 2023-03-14
> So does it make it worse, I don't get it...

It looks to me like it's taking about 1 ms to get the update now, but the old one was taking the time to get the update *and* it was one frame behind, so for 100 Hz, that's 10 ms + the extra time. I think this is how to interpret it? Please let me know if I got something wrong.