@0forks 2023-03-14
> I think this is how to interpret it?

Yes, this should reduce latency in Windows SteamVR by ~8-11ms or more (1 frame) depending on your headset's update rate. Because the bridge thread was waking up only after receiving data, the new tracker positions were stuck until the next frame. This implementation signals the thread when to send too, so it no longer happens.

The previous implementation works and is okay because frame time consistency is more important than latency when presenting frames.

The lower half of the bar chart is the new version, now it takes 1-2ms on average to receive a response by the (simulated) driver on both platforms, if I'm measuring it right. Please do test if there is any perceivable difference.