@0forks 2023-03-14
> we should be inserting explicit waits/getoverlappedresult anywhere we want a blocking-style call

Ah, I just realized you were talking about the part where we read the data, not the waits. Indeed the docs say:

> If hFile is opened with FILE_FLAG_OVERLAPPED, the lpOverlapped parameter must point to a valid and unique [OVERLAPPED](https://learn.microsoft.com/en-us/windows/desktop/api/minwinbase/ns-minwinbase-overlapped) structure, otherwise the function can incorrectly report that the read operation is complete.

As we're peeking and know how many bytes are available, the read operation returns true immediately with data. If not - pipe will close. Maybe we should handle the async case indeed, but JNA doesn't seem to expose `GetOverlappedResult`.