@0forks 2023-03-15
No measurable difference after the last 2 commits:

| Platform / Update rate / Num frames  | 2cc1201 / before (ms) | 9f45d9d / after (ms) |
| ---------------------- | ------------------- | ------------------- |
|                        | `   p1   avg   p99` | `   p1   avg   p99` |
| Windows / 90hz / 5000  | `0.539 1.835 4.110` | `0.543 1.803 4.111` |
| Windows / 100hz / 5000 | `0.536 1.764 4.098` | `0.543 1.825 4.103` |
| Windows / 120hz / 5000 | `0.540 1.806 4.116` | `0.545 1.791 4.108` |

![image](https://user-images.githubusercontent.com/114709761/225424303-7a1c6493-a41e-4da1-a3df-2aaf8de78230.png)