@0forks 2023-03-15
Now comparing to v0.6.3:

| Platform / Update rate / Num frames  | v0.6.3 (ms) | 9f45d9d (ms) |
| ---------------------- | ------------------- | ------------------- |
|                        | `   p1   avg    p99` | `   p1   avg   p99` |
| Windows / 90hz / 5000  | `0.083 9.452 20.208` | `0.546 1.774 4.134` |
| Windows / 100hz / 5000 | `0.071 9.857 19.148` | `0.541 1.773 4.122` |
| Windows / 120hz / 5000 | `0.059 4.196 14.086` | `0.541 1.786 4.115` |

![image](https://user-images.githubusercontent.com/114709761/225439185-3e601417-d169-45eb-b1d7-d3996fd985b2.png)