@kitlith 2023-03-15
> JNA doesn't seem to expose `GetOverlappedResult`.

I ran into that when I was doing my own try at overlapped. This was the approach I took to gain access to `GetOverlappedResult`: https://github.com/kitlith/SlimeVR-Server/blob/unpoll_openvr_thread/server/src/main/java/dev/slimevr/platform/windows/Kernel32.java

Should really get around to submitting an upstream issue.