@0forks 2023-03-17
Looking deeper, it appears that the sleep precision in the VRServer thread is the limiting factor now.
Replacing `Thread.sleep(1)` with `LockSupport.parkNanos(1000000)` improves the latency on Windows to 0.6-1.6ms, but also results in slightly higher CPU usage.

As of now, I believe the latency is already acceptable.

This is how long it sleeps for:
| OS      | Method                         |      p1     |     avg     |     p99     |
|---------|--------------------------------|-------------|-------------|-------------|
| Windows | Thread.sleep(1)                | 1.008800 ms | 2.511393 ms | 4.000000 ms |
| Windows | LockSupport.parkNanos(1000000) | 0.862300 ms | 0.985986 ms | 1.477500 ms |
| Linux   | Thread.sleep(1)                | 1.078178 ms | 1.456425 ms | 1.627606 ms |
| Linux   | LockSupport.parkNanos(1000000) | 1.073479 ms | 1.468993 ms | 1.597515 ms |