@0forks 2023-03-17
Now with VRServer waking on data the p99 latency is under 1ms and it looks smoother under load in VR. Still only properly tested on Windows.