@0forks 2023-03-21
Implemented async IPC and pose updates in the driver, this nuked the remaining 1 frame of latency - body is perfectly in sync with the head now. Will make another PR after cleaning up and more code validation.