@ImUrX 2023-03-21
This seems working nicely, only problem is that apparently the Linux driver has some issues :sob: